separates performance fns, start performance refactor
