+++ shared/common/src/structs/behavior_subject.rs
@@ -1,5 +1,3 @@
-// use futures_util::Stream;
-// use futures_util::StreamExt;
 use std::sync::Arc;
 use tokio::sync::watch::{channel, Receiver, Ref, Sender};
 use tokio_stream::wrappers::WatchStream;
