+++ shared/core/src/performance.rs
@@ -1,43 +1,50 @@
-// use chrono::{Datelike, Local};
-use chrono::{Duration, NaiveDateTime};
-use common::functions::csv::save_csv;
-use common::structs::{Symbol, TradingSettings};
+use chrono::NaiveDateTime;
 use common::{
     constants::DAY_IN_MS,
-    functions::round_down_nth_decimal,
-    // structs::BehaviorSubject,
+    enums::trading_data_update::TradingDataUpdate,
+    functions::{
+        csv::{get_current_env_log_path, save_csv},
+        current_datetime,
+        performance::{
+            calculate_calmar_ratio, calculate_max_drawdown_and_duration,
+            calculate_risk_adjusted_returns, calculate_sharpe_ratio, calculate_sortino_ratio,
+            calculate_success_rate,
+        },
+        round_down_nth_decimal,
+    },
+    structs::{BehaviorSubject, Statistics, Symbol, TradingSettings},
 };
 use glow_error::GlowError;
 use polars::prelude::*;
 use reqwest::Client;
-use std::fmt::Debug;
-
 use std::sync::Mutex;
-use std::{env, fmt};
+use tokio::{spawn, task::JoinHandle};
+use tokio_stream::StreamExt;
 
-#[allow(dead_code)]
 #[derive(Clone)]
 pub struct Performance {
-    http: Client,
+    benchmark_stats: Arc<Mutex<Statistics>>,
+    _http: Client,
+    risk_free_returns: f64,
+    initial_datetime: NaiveDateTime,
+    traded_data_listener: BehaviorSubject<TradingDataUpdate>,
+    trading_stats: Arc<Mutex<Statistics>>,
     trading_settings: Arc<Mutex<TradingSettings>>,
-    trading_initial_datetime: NaiveDateTime,
-    pub benchmark_stats: Statistics,
-    pub trading_stats: Statistics,
-    pub risk_free_returns: f64,
 }
 
 impl Performance {
     pub fn new(
         trading_settings: &Arc<Mutex<TradingSettings>>,
-        trading_initial_datetime: NaiveDateTime,
+        traded_data_listener: &BehaviorSubject<TradingDataUpdate>,
     ) -> Self {
         Self {
-            http: Client::new(),
-            trading_settings: trading_settings.clone(),
-            trading_initial_datetime,
-            benchmark_stats: Statistics::default(),
-            trading_stats: Statistics::default(),
+            benchmark_stats: Arc::new(Mutex::new(Statistics::default())),
+            _http: Client::new(),
             risk_free_returns: 0.0,
+            initial_datetime: current_datetime(),
+            traded_data_listener: traded_data_listener.clone(),
+            trading_stats: Arc::new(Mutex::new(Statistics::default())),
+            trading_settings: trading_settings.clone(),
         }
     }
 
@@ -61,10 +68,11 @@ impl Performance {
 // calmar-ratio = reward / tail risk = CAGR (compound anual growth rate) / max drawdown -> for that matter, use benchmark period growth rate
 
 impl Performance {
-    pub fn set_benchmark_stats(
-        &mut self,
-        benchmark_trading_data: &LazyFrame,
-    ) -> Result<(), GlowError> {
+    fn _set_risk_free_daily_returns(&mut self, returns: f64) {
+        self.risk_free_returns = returns;
+    }
+
+    fn set_benchmark_stats(&self, benchmark_trading_df: DataFrame) -> Result<(), GlowError> {
         let traded_symbol;
         let anchor_symbol;
         {
@@ -73,10 +81,8 @@ impl Performance {
             anchor_symbol = settings.get_anchor_symbol();
         }
 
-        let journey_formmated_datetime_start = self
-            .trading_initial_datetime
-            .format("%H:%M-%d-%m-%Y")
-            .to_string();
+        let journey_formmated_datetime_start =
+            self.initial_datetime.format("%H:%M-%d-%m-%Y").to_string();
 
         let trading_journey_identifier = format!(
             "{}_{}_{}",
@@ -85,16 +91,15 @@ impl Performance {
 
         let path = get_current_env_log_path();
         let file_name = format!("{}_benchmark_data.csv", trading_journey_identifier);
-        let benchmark_trading_data_df = benchmark_trading_data.clone().collect()?;
-        save_csv(path.clone(), file_name, &benchmark_trading_data_df, true)?;
+        save_csv(path.clone(), file_name, &benchmark_trading_df, true)?;
 
-        let (benchmark_data, benchmark_stats) = calculate_benchmark_data(
-            benchmark_trading_data,
-            self.risk_free_returns,
-            traded_symbol,
-        )?;
-
-        self.benchmark_stats = benchmark_stats;
+        let benchmark_trading_lf = benchmark_trading_df.lazy();
+        let (benchmark_data, benchmark_stats) =
+            calculate_benchmark_data(benchmark_trading_lf, self.risk_free_returns, traded_symbol)?;
+        {
+            let mut lock = self.benchmark_stats.lock().unwrap();
+            *lock = benchmark_stats;
+        };
 
         let file_name = format!("{}_benchmark_trades.csv", trading_journey_identifier);
         save_csv(path, file_name, &benchmark_data, true)?;
@@ -102,11 +107,16 @@ impl Performance {
         Ok(())
     }
 
-    fn set_risk_free_daily_returns(&mut self, returns: f64) {
-        self.risk_free_returns = returns;
+    fn handle_initial_traded_data(&self, benchmark_trading_df: DataFrame) {
+        match self.set_benchmark_stats(benchmark_trading_df) {
+            Ok(()) => {}
+            Err(error) => {
+                println!("set_benchmark_stats error {:?}", error);
+            }
+        }
     }
 
-    pub fn update_trading_stats(&mut self, trading_data: &DataFrame) -> Result<(), GlowError> {
+    fn update_trading_stats(&self, traded_data: DataFrame) -> Result<(), GlowError> {
         let traded_symbol;
         let anchor_symbol;
         {
@@ -114,57 +124,72 @@ impl Performance {
             traded_symbol = settings.get_traded_symbol();
             anchor_symbol = settings.get_anchor_symbol();
         }
-
-        let journey_formmated_datetime_start = self
-            .trading_initial_datetime
-            .format("%H:%M-%d-%m-%Y")
-            .to_string();
+        let journey_formmated_datetime_start =
+            self.initial_datetime.format("%H:%M-%d-%m-%Y").to_string();
 
         let trading_journey_identifier = format!(
             "{}_{}_{}",
             journey_formmated_datetime_start, anchor_symbol.name, traded_symbol.name
         );
-
-        let trading_journey_start = self.trading_initial_datetime.timestamp_millis();
-        let filter_mask = trading_data
+        let trading_journey_start = self.initial_datetime.timestamp_millis();
+        let filter_mask = traded_data
             .column("start_time")?
             .gt_eq(trading_journey_start)?;
-        let trading_data = trading_data.filter(&filter_mask)?;
-
+        let trading_data = traded_data.filter(&filter_mask)?;
         let path = get_current_env_log_path();
         let file_name = format!("{}_trading_data.csv", trading_journey_identifier);
-
         save_csv(path.clone(), file_name, &trading_data, true)?;
-
         let (trading_data, trading_stats) = update_trading_data(
             &trading_data,
             self.risk_free_returns,
             traded_symbol,
-            Some(self.trading_initial_datetime),
+            Some(self.initial_datetime),
         )?;
-
-        self.trading_stats = trading_stats;
-
+        {
+            let mut lock = self.trading_stats.lock().unwrap();
+            *lock = trading_stats;
+        };
         let file_name = format!("{}_trades.csv", trading_journey_identifier);
-
         save_csv(path, file_name, &trading_data, true)?;
-
         Ok(())
     }
-}
 
-// TODO: move this from here
-fn get_current_env_log_path() -> String {
-    let env = env::var("ENV_NAME").unwrap().to_uppercase();
-    if env == "PROD".to_string() {
-        "data/journals".to_string()
-    } else {
-        "data/test".to_string()
+    fn handle_market_traded_data(&self, traded_data: DataFrame) {
+        match self.update_trading_stats(traded_data) {
+            Ok(()) => {}
+            Err(error) => {
+                println!("update_trading_stats error {:?}", error);
+            }
+        }
+    }
+
+    fn init_traded_data_handler(&self) -> JoinHandle<()> {
+        let performance = self.clone();
+        spawn(async move {
+            let mut subscription = performance.traded_data_listener.subscribe();
+
+            while let Some(traded_data) = subscription.next().await {
+                match traded_data {
+                    TradingDataUpdate::Initial(benchmark_trading_df) => {
+                        performance.handle_initial_traded_data(benchmark_trading_df)
+                    }
+                    TradingDataUpdate::Market(traded_data) => {
+                        performance.handle_market_traded_data(traded_data)
+                    }
+                    _ => {}
+                }
+            }
+        })
+    }
+
+    pub fn init(self) {
+        self.init_traded_data_handler();
     }
 }
 
+// TODO: move this inside performance impl
 pub fn calculate_benchmark_data(
-    benchmark_trading_data: &LazyFrame,
+    benchmark_trading_data: LazyFrame,
     risk_free_returns: f64,
     traded_symbol: &Symbol,
 ) -> Result<(DataFrame, Statistics), GlowError> {
@@ -189,7 +214,7 @@ pub fn update_trading_data(
     log_from_timestamp_on: Option<NaiveDateTime>,
 ) -> Result<(DataFrame, Statistics), GlowError> {
     let trading_data_lf = trading_data.clone().lazy();
-    let trades_lf = calculate_trades(&trading_data_lf)?;
+    let trades_lf = calculate_trades(trading_data_lf)?;
 
     let mut trading_lf =
         calculate_trading_sessions(trades_lf, traded_symbol, log_from_timestamp_on)?;
@@ -202,8 +227,8 @@ pub fn update_trading_data(
     Ok((df, trading_stats))
 }
 
-pub fn calculate_trades(lf: &LazyFrame) -> Result<LazyFrame, GlowError> {
-    let lf = lf.clone().with_columns([when(
+pub fn calculate_trades(lf: LazyFrame) -> Result<LazyFrame, GlowError> {
+    let lf = lf.with_columns([when(
         col("position")
             .shift(1)
             .is_not_null()
@@ -401,236 +426,6 @@ pub fn calculate_trading_stats(
     ))
 }
 
-fn calculate_success_rate(returns_series: &Series) -> Result<f64, GlowError> {
-    let ca = returns_series.f64()?;
-    let positive_count = ca.into_no_null_iter().filter(|&x| x > 0.0).count();
-    let total_count = ca.into_no_null_iter().len();
-    let success_rate = if total_count != 0 {
-        positive_count as f64 / total_count as f64
-    } else {
-        0.0 as f64
-    };
-
-    Ok(round_down_nth_decimal(success_rate, 2))
-}
-
-fn calculate_risk_adjusted_returns(
-    returns_series: &Series,
-    risk_series: &Series,
-) -> Result<f64, GlowError> {
-    let returns_ca = returns_series.f64()?;
-    let risk_ca = risk_series.f64()?;
-
-    let risk_adjusted_returns = returns_ca
-        .into_no_null_iter()
-        .enumerate()
-        .zip(risk_ca.into_no_null_iter())
-        .fold(0.0, |acc, ((_index, returns), risk)| {
-            let iteration = if risk != 0.0 { returns / risk } else { 0.0 };
-            acc + iteration
-        });
-    Ok(risk_adjusted_returns)
-}
-
-fn calculate_sharpe_ratio(
-    returns_series: &Series,
-    risk_series: &Series,
-    risk_free_returns: f64,
-) -> Result<f64, GlowError> {
-    let returns_ca = returns_series.f64()?;
-    let risk_ca = risk_series.f64()?;
-
-    let risk_adjusted_returns = returns_ca
-        .into_no_null_iter()
-        .zip(risk_ca.into_no_null_iter())
-        .fold(-risk_free_returns, |acc, (returns, risk)| {
-            let iteration = if risk != 0.0 { returns / risk } else { 0.0 };
-            acc + iteration
-        });
-    Ok(risk_adjusted_returns)
-}
-
-fn calculate_sortino_ratio(
-    returns_series: &Series,
-    downside_risk_series: &Series,
-    risk_free_returns: f64,
-) -> Result<f64, GlowError> {
-    let returns_ca = returns_series.f64()?;
-    let downside_risk_ca = downside_risk_series.f64()?;
-
-    let risk_adjusted_returns = returns_ca
-        .into_no_null_iter()
-        .zip(downside_risk_ca.into_no_null_iter())
-        .fold(-risk_free_returns, |acc, (returns, risk)| {
-            let iteration = if risk != 0.0 { returns / risk } else { 0.0 };
-            acc + iteration
-        });
-    Ok(risk_adjusted_returns)
-}
-
-fn calculate_calmar_ratio(balance_series: &Series, max_drawdown: f64) -> Result<f64, GlowError> {
-    if max_drawdown == 0.0 {
-        Ok(0.0)
-    } else {
-        let balances_vec: Vec<f64> = balance_series.f64()?.into_no_null_iter().collect();
-        let last_balance_index = balances_vec.len() - 1;
-        let last_balance = balances_vec[last_balance_index];
-        Ok(last_balance / max_drawdown)
-    }
-}
-
-fn calculate_max_drawdown_and_duration(
-    start_series: &Series,
-    end_series: &Series,
-    drawdown_series: &Series,
-) -> Result<(f64, Duration), GlowError> {
-    let start_vec = start_series
-        .datetime()?
-        .into_no_null_iter()
-        .collect::<Vec<_>>();
-    let end_vec = end_series
-        .datetime()?
-        .into_no_null_iter()
-        .collect::<Vec<_>>();
-    let drawdown_vec = drawdown_series
-        .f64()
-        .unwrap()
-        .into_no_null_iter()
-        .collect::<Vec<f64>>();
-    if let Some((max_index, _)) = drawdown_vec
-        .iter()
-        .enumerate()
-        .max_by(|(_, a), (_, b)| a.partial_cmp(b).unwrap())
-    {
-        let max_drawdown = drawdown_vec[max_index];
-        let max_drawdown_duration =
-            Duration::milliseconds(end_vec[max_index] - start_vec[max_index]);
-
-        Ok((max_drawdown, max_drawdown_duration))
-    } else {
-        Ok((0.0, Duration::minutes(0)))
-    }
-}
-
-// pub fn calculate_performance_on_trading_sessions(lf: LazyFrame) {}
-
-// pub async fn get_latest_us_treasury_bills_yearly_rate(http: &Client) -> Result<f64, Error> {
-//     let now = Local::now();
-//     let first_day_of_year = now.with_month(1).unwrap().with_day(1).unwrap();
-//     let formatted_current_day = now.format("%Y-%m-%d");
-//     let formatted_first_day_of_year = first_day_of_year.format("%Y-%m-%d");
-
-//     let url =  format!("https://api.fiscaldata.treasury.gov/services/api/fiscal_service/v2/accounting/od/avg_interest_rates?fields=record_date,security_desc,avg_interest_rate_amt&filter=record_date:lt:{},src_line_nbr:eq:1,record_date:gt:{}&sort=-record_date", formatted_current_day, formatted_first_day_of_year);
-//     let result: HttpTreasuryResponseWrapper<TreasuryRateResponse> =
-//         http.get(url).send().await?.json().await?;
-
-//     let yearly_rate: f64 = result.data.first().unwrap().avg_interest_rate_amt.parse()?;
-
-//     let current_year = now.year();
-//     let is_leap_year = chrono::NaiveDate::from_ymd_opt(current_year, 1, 1)
-//         .unwrap()
-//         .with_month(12)
-//         .unwrap()
-//         .day()
-//         == 29;
-
-//     let days_in_current_year = if is_leap_year { 366 } else { 365 };
-
-//     let eir_daily =
-//         ((1.0 + yearly_rate) / days_in_current_year as f64).powi(days_in_current_year - 1);
-
-//     Ok(eir_daily)
-// }
-
-#[allow(dead_code)]
-#[derive(Clone, Debug)]
-pub struct Statistics {
-    success_rate: f64,
-    current_balance: f64,
-    risk: f64,
-    downside_deviation: f64,
-    risk_adjusted_return: f64,
-    max_drawdown: f64,
-    max_drawdown_duration: Duration,
-    sharpe_ratio: f64,
-    sortino_ratio: f64,
-    calmar_ratio: f64,
-}
-
-impl fmt::Display for Statistics {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            r#"
-🏆 Success rate (%): {:.2}
-💰 Last balance (USDT): {:.4}
-📊 Risk: {:.6}
-📐 Downside deviation: {:.6}
-📏 Risk adjusted return: {:.6}
-📉 Max drawndown (%): {:.4}
-⏳ Max drawdown duration: {}h{}
-📝 Sharpe: {:.2}
-📝 Sortino: {:.2}
-📝 Calmar: {:.2}"#,
-            self.success_rate,
-            self.current_balance,
-            self.risk,
-            self.downside_deviation,
-            self.risk_adjusted_return,
-            self.max_drawdown,
-            self.max_drawdown_duration.num_hours(),
-            self.max_drawdown_duration.num_minutes() % 60,
-            self.sharpe_ratio,
-            self.sortino_ratio,
-            self.calmar_ratio
-        )
-    }
-}
-
-impl Statistics {
-    fn new(
-        success_rate: f64,
-        current_balance: f64,
-        risk: f64,
-        downside_deviation: f64,
-        risk_adjusted_return: f64,
-        max_drawdown: f64,
-        max_drawdown_duration: Duration,
-        sharpe_ratio: f64,
-        sortino_ratio: f64,
-        calmar_ratio: f64,
-    ) -> Self {
-        Statistics {
-            success_rate,
-            current_balance,
-            risk,
-            downside_deviation,
-            risk_adjusted_return,
-            max_drawdown,
-            max_drawdown_duration,
-            sharpe_ratio,
-            sortino_ratio,
-            calmar_ratio,
-        }
-    }
-}
-
-impl Default for Statistics {
-    fn default() -> Self {
-        Self {
-            success_rate: 0.0,
-            current_balance: 0.0,
-            risk: 0.0,
-            downside_deviation: 0.0,
-            risk_adjusted_return: 0.0,
-            max_drawdown: 0.0,
-            max_drawdown_duration: Duration::minutes(0),
-            sharpe_ratio: 0.0,
-            sortino_ratio: 0.0,
-            calmar_ratio: 0.0,
-        }
-    }
-}
 // #[allow(dead_code)]
 // #[derive(Debug, Deserialize)]
 // pub struct HttpTreasuryResponseWrapper<T> {
